Upgrade to nest 6 Microservice config centralized Meassge client setup simplified
