+++ .vscode/launch.json
@@ -1,0 +1,13 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+      {
+          "type": "node",
+          "request": "attach",
+          "name": "Node: Nodemon",
+          "processId": "${command:PickProcess}",
+          "restart": true,
+          "protocol": "inspector",
+      },
+  ]
+}
