+++ package.json
@@ -20,10 +20,11 @@
     "test:e2e": "jest --config ./test/jest-e2e.json"
   },
   "dependencies": {
-    "@nestjs/common": "^5.4.0",
-    "@nestjs/core": "^5.4.0",
-    "@nestjs/microservices": "^5.7.4",
-    "@nestjs/passport": "^5.1.0",
+    "@nestjs/common": "6.0.2",
+    "@nestjs/core": "6.0.2",
+    "@nestjs/microservices": "6.0.2",
+    "@nestjs/passport": "^6.0.0",
+    "@nestjs/platform-express": "^6.0.2",
     "amqp-connection-manager": "^2.3.0",
     "amqplib": "^0.5.3",
     "class-transformer": "^0.2.0",
@@ -37,7 +38,7 @@
     "typescript": "^3.0.1"
   },
   "devDependencies": {
-    "@nestjs/testing": "^5.1.0",
+    "@nestjs/testing": "6.0.2",
     "@types/express": "^4.16.0",
     "@types/jest": "^23.3.1",
     "@types/node": "^10.7.1",
