--- src/dataApi/app.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { DataApiController } from './dataApi.controller';
-import { DataApiService } from './dataApi.service';
-
-@Module({
-  imports: [],
-  controllers: [DataApiController],
-  providers: [DataApiService],
-})
-export class AppModule {}
