+++ src/dataSteward/dataSteward.controller.ts
@@ -1,12 +1,17 @@
-import { Controller } from '@nestjs/common';
+import { Controller, Inject } from '@nestjs/common';
 import { MessagePattern } from '@nestjs/microservices';
+import { ClientRMQ } from '@nestjs/microservices/client/client-rmq';
 import { DataStewardPersistenceService } from './dataStewardPersistence.service';
 import { Movie } from './dataTypes';
 import { Guid } from 'guid-typescript';
+import { MESSAGE_SERVICE } from 'src/microServiceConfig';
 
 @Controller()
 export class DataStewardController {
-  constructor(private database: DataStewardPersistenceService) {}
+  constructor(
+    @Inject(MESSAGE_SERVICE) private client: ClientRMQ,
+    private database: DataStewardPersistenceService,
+  ) {}
 
   @MessagePattern({ channel: 'createMovie' })
   createMovie(movie: Movie): string {
