+++ src/main.ts
@@ -1,20 +1,17 @@
 import { NestFactory } from '@nestjs/core';
-import { Transport } from '@nestjs/microservices';
-import { AppModule } from './dataApi/app.module';
+import { MicroServiceConfig } from './microServiceConfig';
+import { DataApi } from './dataApi/dataApi.module';
 import { DataSteward } from './dataSteward/dataSteward.module';
 
-async function bootstrap() {
-  const service = await NestFactory.createMicroservice(DataSteward, {
-    transport: Transport.RMQ,
-    options: {
-      urls: [`amqp://localhost:5672`],
-      queue: 'movie_queue',
-      queueOptions: { durable: false },
-    },
-  });
-  service.listen(() => console.log('DataSteward is running.'));
+(async function bootstrap() {
+  const dataSteward = await NestFactory.createMicroservice(
+    DataSteward,
+    MicroServiceConfig,
+  );
+  await dataSteward.listenAsync();
 
-  const app = await NestFactory.create(AppModule);
-  await app.listen(3000);
-}
-bootstrap();
+  const dataApi = await NestFactory.create(DataApi);
+  let dataApiService = dataApi.connectMicroservice(MicroServiceConfig);
+  await dataApiService.listenAsync();
+  await dataApi.listen(3000);
+})();
