+++ test/app.e2e-spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import * as request from 'supertest';
-import { AppModule } from '../src/dataApi/app.module';
+import { AppModule } from '../src/dataApi/dataApi.module';
 
 describe('AppController (e2e)', () => {
   let app;
